Corrgge price

+++ classes/Toy.php
@@ -6,7 +6,7 @@ class Toy extends Category{
     public $format;
     public $price;
 
-    public function __construct(array $typology, array $format, int $price)
+    public function __construct(array $typology, array $format, float $price)
     {
      $this->typology = $typology;
      $this->format = $format;
